GQL-16 refactor(schema): refactor schema
