+++ src/components/Schema/Schema.module.scss
@@ -19,7 +19,7 @@
 .schema {
   position: absolute;
   height: 100%;
-  width: 35%;
+  width: 400px;
   background-color: #f1efef;
   border-left: 5px solid green;
   right: 0;
@@ -30,11 +30,12 @@
 }
 
 .opened {
-  right: 35%;
+  right: 400px;
 }
 
 .loading {
   position: absolute;
   top: 50%;
   left: 50%;
+  transform: translate(-50%, -50%);
 }
