+++ src/index.scss
@@ -1,4 +1,4 @@
-@import "/node_modules/normalize.css";
+@import '/node_modules/normalize.css';
 
 * {
   margin: 0;
@@ -15,7 +15,7 @@ ul {
   display: flex;
   flex-direction: column;
   background-color: #e5e5e5;
-  font-family: "Roboto";
+  font-family: 'Roboto';
 }
 
 main {
